refactor: moved utils/hbCalc.py and utils/huntBotSolver.py to utils/huntbot refactor: utils/configs to utils/config_models           resolved the imports as well fix: included guild in on_channel_switch webhook in `channelswitcher.py` (idk why was it not added, it was supposed to have been fixed) fix: used aiohttp instead of requests for safety_check_loop in `client.py`
