+++ core/client.py
@@ -25,13 +25,12 @@ from datetime import datetime, timedelta, timezone
 import aiohttp
 import discord
 import pytz
-import requests
 from discord.ext import commands, tasks
 
 # Local
 import core.components as components
 import core.database as database
-import utils.configs as config_models
+import utils.config_models as config_models
 import utils.system as syst
 from utils.constants import owo_dusk_api, version
 from utils.loader import (
@@ -256,8 +255,22 @@ class MyClient(commands.Bot):
 
     @tasks.loop(seconds=7)
     async def safety_check_loop(self):
-        safety_check = requests.get(f"{owo_dusk_api}/safety_check.json").json()
-        latest_version = requests.get(f"{owo_dusk_api}/version.json").json()
+        timeout = aiohttp.ClientTimeout(total=10)
+        try:
+            async with self.session.get(
+                f"{owo_dusk_api}/safety_check.json", timeout=timeout
+            ) as resp:
+                safety_check = await resp.json()
+            async with self.session.get(
+                f"{owo_dusk_api}/version.json", timeout=timeout
+            ) as resp:
+                latest_version = await resp.json()
+        except Exception as e:
+            await self.log(
+                f"Safety check request failed, will retry next cycle: {e}",
+                "#5c0018",
+            )
+            return
 
         if syst.system.compare_versions(version, safety_check["version"]):
             self.command_handler_status["captcha"] = True
